requirements.txt
streamlit==1.32.2
mysql-connector-python==8.3.0
textblob==0.17.1
nltk==3.8.1
pandas==2.2.2
